+++ CHANGELOG.md
@@ -41,6 +41,10 @@
 * Terminal content is now auto-scrolled when moving the mouse above or
   below the window while selecting
   (https://codeberg.org/dnkl/foot/issues/149).
+* **font-bold**, **font-italic** and **font-bold-italic** options to
+  `foot.ini`. These options allow custom bold/italic fonts. They are
+  unset by default, meaning the bold/italic version of the regular
+  font is used (https://codeberg.org/dnkl/foot/issues/169).
 
 
 ### Changed